io.c:	in receive_nb changed type of len from int to ssize_t 	changed format in printf to reflect long type ssize_t
common.c:	changed type of avail from int to long to match return value of sizeof

vclient.c:	changed option parsing to use getopt_long
		all option are now working as expected. Order does not matter an longer. You can use multiple -c options on the command line or just add vito commends without -c.

+++ branches/vcontrold_p300_vheat/vcontrold/common.c
@@ -55,7 +55,7 @@ void logIT (int class,char *string) {
 	tPtr=ctime(&t);
 	char dbg[1000];
 	int pid;
-	int avail;
+	long avail;
